Move MessageBodyRetriever to test utils
